\name{indiumoxide}
\alias{indiumoxide}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Diffractogram of Indium oxide doped with tin }
\description{
  Example of a diffractogram used in Davies et al. (2008)
}
\usage{
data(indiumoxide)
}
%- maybe also 'usage' for other objects documented here.
\details{
  The diffractogram was taken on a thin film of \eqn{In_2 O_3 :Sn},
i.e. indium oxide doped with tin. This material is usually called ITO
(indium-tin-oxide). 
It has a good electrical conductivity and is transparent in the visible wave
length region. 
The first column of the matrix contains the angles of diffraction, the second column contains the corresponding photon counts.}
\references{P.L. Davies, U. Gather, M. Meise, D. Mergel, T. Mildenberger (2008): 
"Residual based localization and quantification of peaks in x-ray diffractograms", 
Annals of Applied Statistics, Vol. 2, No. 3, 861-886.. 
\url{http://www.statistik.tu-dortmund.de/fileadmin/user_upload/Lehrstuehle/MSind/Publikationen/2008/2008_-_Davies_Gather_Meise_Mergel_Mildenberger_-_Residual_based_localization_and_quantification_of_peaks_in_x-ray_diffractograms.pdf}
}
\author{ D. Mergel }
%\note{ ~~further notes~~ 
%
% ~Make other sections like Warning with \section{Warning }{....} ~
%}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%}
%
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
